+++ src/sdk/modules/ownableValidator/toOwnableValidator.test.ts
@@ -1,8 +1,4 @@
-import {
-  getAddOwnableExecutorOwnerAction,
-  getExecuteOnOwnedAccountAction,
-  getOwnableValidatorSignature
-} from "@rhinestone/module-sdk"
+import { getOwnableValidatorSignature } from "@rhinestone/module-sdk"
 import {
   http,
   type Account,
@@ -10,19 +6,13 @@ import {
   type Chain,
   type Hex,
   type LocalAccount,
-  type PublicClient,
-  type WalletClient,
   encodeAbiParameters,
   encodeFunctionData,
   encodePacked,
   getAddress,
-  parseAbi,
-  toHex,
   zeroAddress
 } from "viem"
-import { waitForTransactionReceipt } from "viem/actions"
 import { afterAll, beforeAll, describe, expect, test } from "vitest"
-import { testAddresses } from "../../../test/callDatas"
 import { toNetwork } from "../../../test/testSetup"
 import {
   fundAndDeployClients,
@@ -38,138 +28,9 @@ import {
 } from "../../clients/createNexusClient"
 import { parseModuleTypeId } from "../../clients/decorators/erc7579/supportsModule"
 import { K1_VALIDATOR_ADDRESS } from "../../constants"
-import { toK1Validator } from "../k1Validator/toK1Validator"
 import type { Module } from "../utils/Types"
 import { type OwnableActions, ownableActions } from "./decorators"
 import { toOwnableValidator } from "./toOwnableValidator"
-
-describe("modules.ownableValidator.dx", async () => {
-  let network: NetworkConfig
-  let chain: Chain
-  let bundlerUrl: string
-
-  // Test utils
-  let testClient: MasterClient
-  let eoaAccount: LocalAccount
-  let userTwo: LocalAccount
-  let userThree: LocalAccount
-
-  beforeAll(async () => {
-    network = await toNetwork()
-
-    chain = network.chain
-    bundlerUrl = network.bundlerUrl
-    eoaAccount = getTestAccount(0)
-    userTwo = getTestAccount(1)
-    userThree = getTestAccount(2)
-
-    testClient = toTestClient(chain, getTestAccount(5))
-  })
-
-  afterAll(async () => {
-    await killNetwork([network?.rpcPort, network?.bundlerPort])
-  })
-
-  test("should demonstrate ownables module dx", async () => {
-    /**
-     * This test demonstrates the creation and use of an ownables module for multi-signature functionality:
-     *
-     * 1. Setup and Installation:
-     *    - Create a Nexus client for the main account
-     *    - Install the ownables module on the smart contract account
-     *
-     * 2. Multi-Signature Transaction:
-     *    - Prepare a user operation (withdrawal) that requires multiple signatures
-     *    - Collect signatures from required owners
-     *    - Execute the multi-sig transaction
-     *
-     * This test showcases how the ownables module enables multi-signature functionality
-     * on a smart contract account, ensuring that certain actions require approval from
-     * multiple designated owners.
-     */
-
-    // Create a Nexus client for the main account (eoaAccount)
-    // This client will be used to interact with the smart contract account
-    const nexusClient = await createNexusClient({
-      signer: eoaAccount,
-      chain,
-      transport: http(),
-      bundlerTransport: http(bundlerUrl)
-    })
-
-    // Fund the account and deploy the smart contract wallet
-    // This is just a reminder to fund the account and deploy the smart contract wallet
-    await fundAndDeployClients(testClient, [nexusClient])
-
-    // Create an ownables module with the following configuration:
-    // - Threshold: 2 (requires 2 signatures for approval)
-    // - Owners: userThree and userTwo
-    const ownableModule = toOwnableValidator({
-      account: nexusClient.account,
-      signer: eoaAccount,
-      moduleInitArgs: {
-        threshold: 2n,
-        owners: [userThree.address, userTwo.address]
-      }
-    })
-
-    // Install the ownables module on the Nexus client's smart contract account
-    const hash = await nexusClient.installModule({
-      module: ownableModule.moduleInitData
-    })
-
-    // Extend the Nexus client with ownable-specific actions
-    // This allows the client to use the new module's functionality
-    const ownableNexusClient = nexusClient.extend(ownableActions(ownableModule))
-
-    // Wait for the module installation transaction to be mined and check its success
-    const { success } = await ownableNexusClient.waitForUserOperationReceipt({
-      hash
-    })
-
-    // Prepare a user operation to withdraw 1 wei to userTwo
-    // This demonstrates a simple transaction that requires multi-sig approval
-    // @ts-ignore
-    const withdrawalUserOp = await ownableNexusClient.prepareUserOperation({
-      calls: [
-        {
-          to: userTwo.address,
-          value: 1n
-        }
-      ]
-    })
-
-    // Get the hash of the user operation
-    // This hash will be signed by the required owners
-    const withdrawalUserOpHash =
-      // @ts-ignore
-      await nexusClient.account.getUserOpHash(withdrawalUserOp)
-
-    // Collect signatures from both required owners (userTwo and userThree)
-    const signatures = await Promise.all(
-      [userTwo, userThree].map(async (signer) => {
-        return signer.signMessage({
-          message: { raw: withdrawalUserOpHash }
-        })
-      })
-    )
-
-    // Combine the signatures and set them on the user operation
-    // The order of signatures should match the order of owners in the module configuration
-    withdrawalUserOp.signature = await ownableNexusClient.prepareSignatures({
-      signatures
-    })
-    // Send the user operation with the collected signatures
-    const userOpHash = await nexusClient.sendUserOperation(withdrawalUserOp)
-
-    // Wait for the user operation to be mined and check its success
-    const { success: userOpSuccess } =
-      await ownableNexusClient.waitForUserOperationReceipt({ hash: userOpHash })
-
-    // Verify that the multi-sig transaction was successful
-    expect(userOpSuccess).toBe(true)
-  })
-})
 
 describe("modules.ownables", async () => {
   let network: NetworkConfig
@@ -439,143 +300,5 @@ describe("modules.ownables", async () => {
     const [installedValidatorsAfter] =
       await nexusClient.getInstalledValidators()
     expect(installedValidatorsAfter).toEqual([K1_VALIDATOR_ADDRESS])
-  })
-})
-
-describe("modules.ownableExecutor", async () => {
-  let network: NetworkConfig
-  let chain: Chain
-  let bundlerUrl: string
-
-  // Test utils
-  let testClient: MasterClient
-  let eoaAccount: Account
-  let nexusClient: NexusClient
-  let nexusAccountAddress: Address
-  let recipient: Account
-  let recipientAddress: Address
-  let k1Module: Module
-  beforeAll(async () => {
-    network = await toNetwork()
-
-    chain = network.chain
-    bundlerUrl = network.bundlerUrl
-    eoaAccount = getTestAccount(0)
-    recipient = getTestAccount(1)
-    recipientAddress = recipient.address
-
-    testClient = toTestClient(chain, getTestAccount(5))
-
-    nexusClient = await createNexusClient({
-      signer: eoaAccount,
-      chain,
-      transport: http(),
-      bundlerTransport: http(bundlerUrl)
-    })
-
-    nexusAccountAddress = await nexusClient.account.getCounterFactualAddress()
-    await fundAndDeployClients(testClient, [nexusClient])
-
-    const k1Module = toK1Validator({
-      signer: eoaAccount,
-      accountAddress: nexusClient.account.address
-    })
-
-    nexusClient.account.setModule(k1Module)
-  })
-
-  afterAll(async () => {
-    await killNetwork([network?.rpcPort, network?.bundlerPort])
-  })
-
-  test("should install OwnableExecutor module", async () => {
-    const isInstalled = await nexusClient.isModuleInstalled({
-      module: {
-        type: "executor",
-        address: testAddresses.OwnableExecutor
-      }
-    })
-    expect(isInstalled).toBe(false)
-
-    const userOpHash = await nexusClient.installModule({
-      module: {
-        type: "executor",
-        address: testAddresses.OwnableExecutor,
-        initData: encodePacked(["address"], [eoaAccount.address])
-      }
-    })
-    expect(userOpHash).toBeDefined()
-    const receipt = await nexusClient.waitForUserOperationReceipt({
-      hash: userOpHash
-    })
-    expect(receipt.success).toBe(true)
-
-    const isInstalledAfter = await nexusClient.isModuleInstalled({
-      module: {
-        type: "executor",
-        address: testAddresses.OwnableExecutor
-      }
-    })
-    expect(isInstalledAfter).toBe(true)
-  })
-
-  test("should add another EOA as executor", async () => {
-    const execution = await getAddOwnableExecutorOwnerAction({
-      owner: recipientAddress,
-      client: nexusClient.account.client as PublicClient,
-      account: {
-        address: nexusClient.account.address,
-        type: "nexus",
-        deployedOnChains: []
-      }
-    })
-    const userOpHash = await nexusClient.sendTransaction({
-      calls: [
-        {
-          to: testAddresses.OwnableExecutor,
-          data: execution.callData,
-          value: 0n
-        }
-      ]
-    })
-    expect(userOpHash).toBeDefined()
-    const masterClient = nexusClient.account.client as MasterClient
-    const owners = await masterClient.readContract({
-      address: testAddresses.OwnableExecutor,
-      abi: parseAbi([
-        "function getOwners(address account) external view returns (address[])"
-      ]),
-      functionName: "getOwners",
-      args: [nexusClient.account.address]
-    })
-    expect(owners).toContain(recipientAddress)
-  })
-
-  test("added executor EOA should execute user operation on smart account", async () => {
-    const executeOnOwnedAccountExecution = getExecuteOnOwnedAccountAction({
-      execution: {
-        target: zeroAddress,
-        callData: toHex("0x"),
-        value: 0n,
-        to: zeroAddress,
-        data: "0x"
-      },
-      ownedAccount: nexusClient.account.address
-    })
-
-    const client = nexusClient.account.client as WalletClient
-    const hash = await client.sendTransaction({
-      account: recipient,
-      to: testAddresses.OwnableExecutor,
-      data: executeOnOwnedAccountExecution.callData,
-      chain,
-      value: 0n
-    })
-
-    const receipt = await waitForTransactionReceipt(
-      nexusClient.account.client as PublicClient,
-      { hash }
-    )
-    expect(receipt.status).toBe("success")
   })
 })
